third_party/WebKit/LayoutTests/http/tests/worklet/webexposed/global-interface-listing-paint-worklet-expected.txt

CONSOLE MESSAGE: line 139: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE: line 139:     method constructor
CONSOLE MESSAGE: line 139:     method size
CONSOLE MESSAGE: line 139: interface CountQueuingStrategy
CONSOLE MESSAGE: line 139:     method constructor
CONSOLE MESSAGE: line 139:     method size
CONSOLE MESSAGE: line 139: interface DataView
CONSOLE MESSAGE: line 139:     getter buffer
CONSOLE MESSAGE: line 139:     getter byteLength
CONSOLE MESSAGE: line 139:     getter byteOffset
CONSOLE MESSAGE: line 139:     method constructor
CONSOLE MESSAGE: line 139:     method getFloat32
CONSOLE MESSAGE: line 139:     method getFloat64
CONSOLE MESSAGE: line 139:     method getInt16
CONSOLE MESSAGE: line 139:     method getInt32
CONSOLE MESSAGE: line 139:     method getInt8
CONSOLE MESSAGE: line 139:     method getUint16
CONSOLE MESSAGE: line 139:     method getUint32
CONSOLE MESSAGE: line 139:     method getUint8
CONSOLE MESSAGE: line 139:     method setFloat32
CONSOLE MESSAGE: line 139:     method setFloat64
CONSOLE MESSAGE: line 139:     method setInt16
CONSOLE MESSAGE: line 139:     method setInt32
CONSOLE MESSAGE: line 139:     method setInt8
CONSOLE MESSAGE: line 139:     method setUint16
CONSOLE MESSAGE: line 139:     method setUint32
CONSOLE MESSAGE: line 139:     method setUint8
CONSOLE MESSAGE: line 139: interface PaintRenderingContext2D
CONSOLE MESSAGE: line 139:     getter currentTransform
CONSOLE MESSAGE: line 139:     getter fillStyle
CONSOLE MESSAGE: line 139:     getter globalAlpha
CONSOLE MESSAGE: line 139:     getter globalCompositeOperation
CONSOLE MESSAGE: line 139:     getter imageSmoothingEnabled
CONSOLE MESSAGE: line 139:     getter imageSmoothingQuality
CONSOLE MESSAGE: line 139:     getter lineCap
CONSOLE MESSAGE: line 139:     getter lineDashOffset
CONSOLE MESSAGE: line 139:     getter lineJoin
CONSOLE MESSAGE: line 139:     getter lineWidth
CONSOLE MESSAGE: line 139:     getter miterLimit
CONSOLE MESSAGE: line 139:     getter shadowBlur
CONSOLE MESSAGE: line 139:     getter shadowColor
CONSOLE MESSAGE: line 139:     getter shadowOffsetX
CONSOLE MESSAGE: line 139:     getter shadowOffsetY
CONSOLE MESSAGE: line 139:     getter strokeStyle
CONSOLE MESSAGE: line 139:     method arc
CONSOLE MESSAGE: line 139:     method arcTo
CONSOLE MESSAGE: line 139:     method beginPath
CONSOLE MESSAGE: line 139:     method bezierCurveTo
CONSOLE MESSAGE: line 139:     method clearRect
CONSOLE MESSAGE: line 139:     method clip
CONSOLE MESSAGE: line 139:     method closePath
CONSOLE MESSAGE: line 139:     method constructor
CONSOLE MESSAGE: line 139:     method createLinearGradient
CONSOLE MESSAGE: line 139:     method createPattern
CONSOLE MESSAGE: line 139:     method createRadialGradient
CONSOLE MESSAGE: line 139:     method drawImage
CONSOLE MESSAGE: line 139:     method ellipse
CONSOLE MESSAGE: line 139:     method fill
CONSOLE MESSAGE: line 139:     method fillRect
CONSOLE MESSAGE: line 139:     method getLineDash
CONSOLE MESSAGE: line 139:     method isPointInPath
CONSOLE MESSAGE: line 139:     method isPointInStroke
CONSOLE MESSAGE: line 139:     method lineTo
CONSOLE MESSAGE: line 139:     method moveTo
CONSOLE MESSAGE: line 139:     method quadraticCurveTo
CONSOLE MESSAGE: line 139:     method rect
CONSOLE MESSAGE: line 139:     method resetTransform
CONSOLE MESSAGE: line 139:     method restore
CONSOLE MESSAGE: line 139:     method rotate
CONSOLE MESSAGE: line 139:     method save
CONSOLE MESSAGE: line 139:     method scale
CONSOLE MESSAGE: line 139:     method setLineDash
CONSOLE MESSAGE: line 139:     method setTransform
CONSOLE MESSAGE: line 139:     method stroke
CONSOLE MESSAGE: line 139:     method strokeRect
CONSOLE MESSAGE: line 139:     method transform
CONSOLE MESSAGE: line 139:     method translate
CONSOLE MESSAGE: line 139:     setter currentTransform
CONSOLE MESSAGE: line 139:     setter fillStyle
CONSOLE MESSAGE: line 139:     setter globalAlpha
CONSOLE MESSAGE: line 139:     setter globalCompositeOperation
CONSOLE MESSAGE: line 139:     setter imageSmoothingEnabled
CONSOLE MESSAGE: line 139:     setter imageSmoothingQuality
CONSOLE MESSAGE: line 139:     setter lineCap
CONSOLE MESSAGE: line 139:     setter lineDashOffset
CONSOLE MESSAGE: line 139:     setter lineJoin
CONSOLE MESSAGE: line 139:     setter lineWidth
CONSOLE MESSAGE: line 139:     setter miterLimit
CONSOLE MESSAGE: line 139:     setter shadowBlur
CONSOLE MESSAGE: line 139:     setter shadowColor
CONSOLE MESSAGE: line 139:     setter shadowOffsetX
CONSOLE MESSAGE: line 139:     setter shadowOffsetY
CONSOLE MESSAGE: line 139:     setter strokeStyle
CONSOLE MESSAGE: line 139: interface PaintWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE: line 139:     method constructor
CONSOLE MESSAGE: line 139: interface ReadableStream
CONSOLE MESSAGE: line 139:     getter locked
CONSOLE MESSAGE: line 139:     method cancel
CONSOLE MESSAGE: line 139:     method constructor
CONSOLE MESSAGE: line 139:     method getReader
CONSOLE MESSAGE: line 139:     method tee
CONSOLE MESSAGE: line 139: interface WorkletGlobalScope
CONSOLE MESSAGE: line 139:     getter console
CONSOLE MESSAGE: line 139:     method constructor
CONSOLE MESSAGE: line 139:     setter console
CONSOLE MESSAGE: line 139: global object
CONSOLE MESSAGE: line 139:     method gc
CONSOLE MESSAGE: line 139:     method registerPaint
